pyproject.toml
[tool.poetry]
name = "graduatework-authservice"
version = "0.1.0"
description = ""
authors = ["sergeyivashkevich <e-box@live.ru>"]
readme = "README.md"
packages = [{include = "graduatework_authservice"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
